backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
hdbcli
python-dotenv
pydantic
